refactor: change dir system
