+++ src/test/kotlin/ecommerce/controller/guest/AuthControllerTest.kt
@@ -1,4 +1,4 @@
-package ecommerce.controller
+package ecommerce.controller.guest
 
 import ecommerce.dto.auth.TokenRequest
 import ecommerce.dto.user.UserDTO
@@ -6,7 +6,7 @@ import ecommerce.enums.UserRole
 import ecommerce.service.AuthService
 import io.restassured.RestAssured
 import io.restassured.http.ContentType
-import org.assertj.core.api.Assertions.assertThat
+import org.assertj.core.api.Assertions
 import org.junit.jupiter.api.Test
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.boot.test.context.SpringBootTest
@@ -32,8 +32,8 @@ class AuthControllerTest {
                 .contentType(ContentType.JSON)
                 .`when`().post("api/auth/signup")
                 .then().log().all().extract()
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.CREATED.value())
-        assertThat(response.header("Authorization")).isNotNull
+        Assertions.assertThat(response.statusCode()).isEqualTo(HttpStatus.CREATED.value())
+        Assertions.assertThat(response.header("Authorization")).isNotNull
     }
 
     @Test
@@ -52,8 +52,8 @@ class AuthControllerTest {
                 .contentType(ContentType.JSON)
                 .`when`().post("api/auth/signup")
                 .then().log().all().extract()
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.CREATED.value())
-        assertThat(response.header("Authorization")).isNotNull
+        Assertions.assertThat(response.statusCode()).isEqualTo(HttpStatus.CREATED.value())
+        Assertions.assertThat(response.header("Authorization")).isNotNull
     }
 
     @Test
@@ -78,7 +78,7 @@ class AuthControllerTest {
                 .`when`().post("api/auth/signIn")
                 .then().log().all().extract()
 
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value())
-        assertThat(response.header("Authorization")).isNotEmpty
+        Assertions.assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value())
+        Assertions.assertThat(response.header("Authorization")).isNotEmpty
     }
 }
